+++ tests/conftest.py
@@ -29,3 +29,15 @@ def fake_data(tmpdir):
 def today_is_00(monkeypatch):
     monkeypatch.setattr('dothebackup.tools.today',
                         lambda: '00')
+
+
+@pytest.fixture
+def something_to_commit(monkeypatch):
+    monkeypatch.setattr('dothebackup.tools.subprocess.check_output',
+                        lambda path, shell: b'A  foo/bar.py')
+
+
+@pytest.fixture
+def nothing_to_commit(monkeypatch):
+    monkeypatch.setattr('dothebackup.tools.subprocess.check_output',
+                        lambda path, shell: b'')